requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20241021
Django==4.2.7
django-crispy-forms==2.0
django-import-export==3.3.0
djangorestframework==3.14.0
et_xmlfile==2.0.0
idna==3.11
MarkupPy==1.18
odfpy==1.4.1
openpyxl==3.1.5
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
requests==2.31.0
sqlparse==0.5.4
tablib==3.5.0
tzdata==2025.2
urllib3==2.6.1
xlrd==2.0.2
xlwt==1.3.0
googlemaps==4.10.0
